refactor(cli): move the watcher's serialized runner into @internals/utils
createBuildRunner had nothing watcher-specific, so it moves to
internals/utils/src/promise.ts as createSerialRunner taking a single
options object, with its tests. The watcher in the generate runner now
imports it.
