+++ internals/utils/src/index.ts
@@ -8,7 +8,7 @@ export { canUseTTY, isCIEnvironment } from './env.ts'
 export { BuildError, getErrorMessage, toCause, toError } from './errors.ts'
 export { clean, exists, findPackageJSON, getRelativePath, read, toFilePath, toPosixPath, trimExtName, write } from './fs.ts'
 export type { PossiblePromise, Streamable } from './promise.ts'
-export { arrayToAsyncIterable, isPromise, memoize } from './promise.ts'
+export { arrayToAsyncIterable, createSerialRunner, isPromise, memoize } from './promise.ts'
 export { isIdentifier, isValidVarName, transformReservedWord } from './reserved.ts'
 export { runtime } from './runtime.ts'
 export { singleQuote } from './string.ts'
