+++ packages/cli/src/runners/generate/utils.test.ts
@@ -1,8 +1,8 @@
 import process from 'node:process'
 import { AsyncEventEmitter } from '@internals/utils'
 import type { KubbHooks } from '@kubb/core'
-import { describe, expect, it, vi } from 'vitest'
-import { createBuildRunner, isNewerVersion, runHook } from './utils.ts'
+import { describe, expect, it } from 'vitest'
+import { isNewerVersion, runHook } from './utils.ts'
 
 const node = process.execPath
 
@@ -138,51 +138,5 @@ describe('isNewerVersion', () => {
 
   it('returns false for a malformed latest version', () => {
     expect(isNewerVersion('5.9.0', 'not-a-version')).toBe(false)
-  })
-})
-
-describe('createBuildRunner', () => {
-  it('collapses triggers that land during a build into one rerun', async () => {
-    let calls = 0
-    const gates: Array<() => void> = []
-    const runner = createBuildRunner(
-      () =>
-        new Promise<void>((resolve) => {
-          calls += 1
-          gates.push(resolve)
-        }),
-      () => {},
-    )
-
-    const first = runner()
-    void runner()
-    void runner()
-    void runner()
-    expect(calls).toBe(1)
-
-    gates[0]?.()
-    await vi.waitFor(() => expect(calls).toBe(2))
-
-    gates[1]?.()
-    await first
-    expect(calls).toBe(2)
-  })
-
-  it('reports a build error through onError and keeps accepting triggers', async () => {
-    const errors: Array<string> = []
-    let shouldFail = true
-    const runner = createBuildRunner(
-      async () => {
-        if (shouldFail) throw new Error('build exploded')
-      },
-      (error) => errors.push(error.message),
-    )
-
-    await runner()
-    expect(errors).toStrictEqual(['build exploded'])
-
-    shouldFail = false
-    await runner()
-    expect(errors).toStrictEqual(['build exploded'])
   })
 })
